Comments, Comments, Comments
NOTE: FROM HERE ON OUT, ANY CHANGES THAT ARE MADE TO THE FILES SHOULD BE
DOCUMENTED IN THE CHANGE LOG, IT IS CURRENTLY UP TO DATE WITH THE MOST
RECENT COMMIT. Added all the headers to each file, with change logs.
Also detailed the purpose of each file. If people want to go in an edit
them that's fine.